@@ -4,7 +4,7 @@ version: 0.1.0
 homepage: https://github.com/pzentenoe/flutter_card_swipper
 
 environment:
-  sdk: '>=2.11.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
