@@ -2,34 +2,34 @@ part of '../src/swiper.dart';
 
 abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
     with SingleTickerProviderStateMixin {
-  double _swiperWidth;
-  double _swiperHeight;
-  Animation<double> _animation;
-  AnimationController _animationController;
-  int _startIndex;
-  int _animationCount;
+  double? _swiperWidth;
+  double? _swiperHeight;
+  late Animation<double> _animation;
+  AnimationController? _animationController;
+  late int _startIndex;
+  int? _animationCount;
 
   @override
   void initState() {
     if (widget.itemWidth == null) {
-      throw new Exception(
-          "==============\n\nwidget.itemWith must not be null when use stack layout.\n========\n");
+      throw Exception(
+          '==============\n\nwidget.itemWith must not be null when use stack layout.\n========\n');
     }
 
     _createAnimationController();
-    widget.controller.addListener(_onController);
+    widget.controller!.addListener(_onController);
     super.initState();
   }
 
   void _createAnimationController() {
-    _animationController = new AnimationController(vsync: this, value: 0.5);
-    Tween<double> tween = new Tween(begin: 0.0, end: 1.0);
-    _animation = tween.animate(_animationController);
+    _animationController = AnimationController(vsync: this, value: 0.5);
+    Tween<double> tween = Tween(begin: 0.0, end: 1.0);
+    _animation = tween.animate(_animationController!);
   }
 
   @override
   void didChangeDependencies() {
-    WidgetsBinding.instance.addPostFrameCallback(_getSize);
+    WidgetsBinding.instance!.addPostFrameCallback(_getSize);
     super.didChangeDependencies();
   }
 
@@ -39,7 +39,7 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
 
   @mustCallSuper
   void afterRender() {
-    RenderObject renderObject = context.findRenderObject();
+    final renderObject = context.findRenderObject()!;
     Size size = renderObject.paintBounds.size;
     _swiperWidth = size.width;
     _swiperHeight = size.height;
@@ -49,12 +49,12 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
   @override
   void didUpdateWidget(T oldWidget) {
     if (widget.controller != oldWidget.controller) {
-      oldWidget.controller.removeListener(_onController);
-      widget.controller.addListener(_onController);
+      oldWidget.controller!.removeListener(_onController);
+      widget.controller!.addListener(_onController);
     }
 
     if (widget.loop != oldWidget.loop) {
-      if (!widget.loop) {
+      if (!widget.loop!) {
         _currentIndex = _ensureIndex(_currentIndex);
       }
     }
@@ -63,16 +63,16 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
   }
 
   int _ensureIndex(int index) {
-    index = index % widget.itemCount;
+    index = index % widget.itemCount!;
     if (index < 0) {
-      index += widget.itemCount;
+      index += widget.itemCount!;
     }
     return index;
   }
 
   @override
   void dispose() {
-    widget.controller.removeListener(_onController);
+    widget.controller!.removeListener(_onController);
     _animationController?.dispose();
     super.dispose();
   }
@@ -80,33 +80,33 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
   Widget _buildItem(int i, int realIndex, double animationValue);
 
   Widget _buildContainer(List<Widget> list) {
-    return new Stack(
+    return Stack(
       children: list,
     );
   }
 
-  Widget _buildAnimation(BuildContext context, Widget w) {
+  Widget _buildAnimation(BuildContext context, Widget? w) {
     List<Widget> list = [];
 
     double animationValue = _animation.value;
 
-    for (int i = 0; i < _animationCount; ++i) {
+    for (int i = 0; i < _animationCount!; ++i) {
       int realIndex = _currentIndex + i + _startIndex;
-      realIndex = realIndex % widget.itemCount;
+      realIndex = realIndex % widget.itemCount!;
       if (realIndex < 0) {
-        realIndex += widget.itemCount;
+        realIndex += widget.itemCount!;
       }
 
       list.add(_buildItem(i, realIndex, animationValue));
     }
 
-    return new GestureDetector(
+    return GestureDetector(
       behavior: HitTestBehavior.opaque,
       onPanStart: _onPanStart,
       onPanEnd: _onPanEnd,
       onPanUpdate: _onPanUpdate,
-      child: new ClipRect(
-        child: new Center(
+      child: ClipRect(
+        child: Center(
           child: _buildContainer(list),
         ),
       ),
@@ -116,33 +116,33 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
   @override
   Widget build(BuildContext context) {
     if (_animationCount == null) {
-      return new Container();
+      return Container();
     }
-    return new AnimatedBuilder(
-        animation: _animationController, builder: _buildAnimation);
+    return AnimatedBuilder(
+        animation: _animationController!, builder: _buildAnimation);
   }
 
-  double _currentValue;
-  double _currentPos;
+  late double _currentValue;
+  late double _currentPos;
 
   bool _lockScroll = false;
 
-  void _move(double position, {int nextIndex}) async {
+  void _move(double position, {int? nextIndex}) async {
     if (_lockScroll) return;
     try {
       _lockScroll = true;
-      await _animationController.animateTo(position,
-          duration: new Duration(milliseconds: widget.duration),
-          curve: widget.curve);
+      await _animationController!.animateTo(position,
+          duration: Duration(milliseconds: widget.duration!),
+          curve: widget.curve!);
       if (nextIndex != null) {
-        widget.onIndexChanged(widget.getCorrectIndex(nextIndex));
+        widget.onIndexChanged!(widget.getCorrectIndex(nextIndex));
       }
     } catch (e) {
       print(e);
     } finally {
       if (nextIndex != null) {
         try {
-          _animationController.value = 0.5;
+          _animationController!.value = 0.5;
         } catch (e) {
           print(e);
         }
@@ -154,36 +154,36 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
   }
 
   int _nextIndex() {
-    int index = _currentIndex + 1;
-    if (!widget.loop && index >= widget.itemCount - 1) {
-      return widget.itemCount - 1;
+    final index = _currentIndex + 1;
+    if (!widget.loop! && index >= widget.itemCount! - 1) {
+      return widget.itemCount! - 1;
     }
     return index;
   }
 
   int _prevIndex() {
-    int index = _currentIndex - 1;
-    if (!widget.loop && index < 0) {
+    final int index = _currentIndex - 1;
+    if (!widget.loop! && index < 0) {
       return 0;
     }
     return index;
   }
 
   void _onController() {
-    switch (widget.controller.event) {
+    switch (widget.controller!.event) {
       case IndexController.PREVIOUS:
-        int prevIndex = _prevIndex();
+        final int prevIndex = _prevIndex();
         if (prevIndex == _currentIndex) return;
         _move(1.0, nextIndex: prevIndex);
         break;
       case IndexController.NEXT:
-        int nextIndex = _nextIndex();
+        final int nextIndex = _nextIndex();
         if (nextIndex == _currentIndex) return;
         _move(0.0, nextIndex: nextIndex);
         break;
       case IndexController.MOVE:
-        throw new Exception(
-            "Custom layout does not support SwiperControllerEvent.MOVE_INDEX yet!");
+        throw Exception(
+            'Custom layout does not support SwiperControllerEvent.MOVE_INDEX yet!');
       case SwiperController.STOP_AUTOPLAY:
       case SwiperController.START_AUTOPLAY:
         break;
@@ -193,17 +193,17 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
   void _onPanEnd(DragEndDetails details) {
     if (_lockScroll) return;
 
-    double velocity = widget.scrollDirection == Axis.horizontal
+    final double velocity = widget.scrollDirection == Axis.horizontal
         ? details.velocity.pixelsPerSecond.dx
         : details.velocity.pixelsPerSecond.dy;
 
-    if (_animationController.value >= 0.75 || velocity > 500.0) {
-      if (_currentIndex <= 0 && !widget.loop) {
+    if (_animationController!.value >= 0.75 || velocity > 500.0) {
+      if (_currentIndex <= 0 && !widget.loop!) {
         return;
       }
       _move(1.0, nextIndex: _currentIndex - 1);
-    } else if (_animationController.value < 0.25 || velocity < -500.0) {
-      if (_currentIndex >= widget.itemCount - 1 && !widget.loop) {
+    } else if (_animationController!.value < 0.25 || velocity < -500.0) {
+      if (_currentIndex >= widget.itemCount! - 1 && !widget.loop!) {
         return;
       }
       _move(0.0, nextIndex: _currentIndex + 1);
@@ -214,7 +214,7 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
 
   void _onPanStart(DragStartDetails details) {
     if (_lockScroll) return;
-    _currentValue = _animationController.value;
+    _currentValue = _animationController!.value;
     _currentPos = widget.scrollDirection == Axis.horizontal
         ? details.globalPosition.dx
         : details.globalPosition.dy;
@@ -227,11 +227,11 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
                     ? details.globalPosition.dx
                     : details.globalPosition.dy) -
                 _currentPos) /
-            _swiperWidth /
+            _swiperWidth! /
             2;
     // no loop ?
-    if (!widget.loop) {
-      if (_currentIndex >= widget.itemCount - 1) {
+    if (!widget.loop!) {
+      if (_currentIndex >= widget.itemCount! - 1) {
         if (value < 0.5) {
           value = 0.5;
         }
@@ -242,28 +242,28 @@ abstract class _CustomLayoutStateBase<T extends _SubSwiper> extends State<T>
       }
     }
 
-    _animationController.value = value;
+    _animationController!.value = value;
   }
 
   int _currentIndex = 0;
 }
 
-double _getValue(List<double> values, double animationValue, int index) {
-  double s = values[index];
+double? _getValue(List<double?> values, double animationValue, int index) {
+  double? s = values[index];
   if (animationValue >= 0.5) {
     if (index < values.length - 1) {
-      s = s + (values[index + 1] - s) * (animationValue - 0.5) * 2.0;
+      s = s! + (values[index + 1]! - s) * (animationValue - 0.5) * 2.0;
     }
   } else {
     if (index != 0) {
-      s = s - (s - values[index - 1]) * (0.5 - animationValue) * 2.0;
+      s = s! - (s - values[index - 1]!) * (0.5 - animationValue) * 2.0;
     }
   }
   return s;
 }
 
 Offset _getOffsetValue(List<Offset> values, double animationValue, int index) {
-  Offset s = values[index];
+  final Offset s = values[index];
   double dx = s.dx;
   double dy = s.dy;
   if (animationValue >= 0.5) {
@@ -277,11 +277,11 @@ Offset _getOffsetValue(List<Offset> values, double animationValue, int index) {
       dy = dy - (dy - values[index - 1].dy) * (0.5 - animationValue) * 2.0;
     }
   }
-  return new Offset(dx, dy);
+  return Offset(dx, dy);
 }
 
 abstract class TransformBuilder<T> {
-  List<T> values;
+  List<T>? values;
 
   TransformBuilder({this.values});
 
@@ -291,21 +291,22 @@ abstract class TransformBuilder<T> {
 class ScaleTransformBuilder extends TransformBuilder<double> {
   final Alignment alignment;
 
-  ScaleTransformBuilder({List<double> values, this.alignment: Alignment.center})
+  ScaleTransformBuilder(
+      {List<double>? values, this.alignment: Alignment.center})
       : super(values: values);
 
   Widget build(int i, double animationValue, Widget widget) {
-    double s = _getValue(values, animationValue, i);
-    return new Transform.scale(scale: s, child: widget);
+    final double s = _getValue(values!, animationValue, i)!;
+    return Transform.scale(scale: s, child: widget);
   }
 }
 
 class OpacityTransformBuilder extends TransformBuilder<double> {
-  OpacityTransformBuilder({List<double> values}) : super(values: values);
+  OpacityTransformBuilder({List<double>? values}) : super(values: values);
 
   Widget build(int i, double animationValue, Widget widget) {
-    double v = _getValue(values, animationValue, i);
-    return new Opacity(
+    final double v = _getValue(values!, animationValue, i)!;
+    return Opacity(
       opacity: v,
       child: widget,
     );
@@ -313,11 +314,11 @@ class OpacityTransformBuilder extends TransformBuilder<double> {
 }
 
 class RotateTransformBuilder extends TransformBuilder<double> {
-  RotateTransformBuilder({List<double> values}) : super(values: values);
+  RotateTransformBuilder({List<double>? values}) : super(values: values);
 
   Widget build(int i, double animationValue, Widget widget) {
-    double v = _getValue(values, animationValue, i);
-    return new Transform.rotate(
+    final double v = _getValue(values!, animationValue, i)!;
+    return Transform.rotate(
       angle: v,
       child: widget,
     );
@@ -325,12 +326,12 @@ class RotateTransformBuilder extends TransformBuilder<double> {
 }
 
 class TranslateTransformBuilder extends TransformBuilder<Offset> {
-  TranslateTransformBuilder({List<Offset> values}) : super(values: values);
+  TranslateTransformBuilder({List<Offset>? values}) : super(values: values);
 
   @override
   Widget build(int i, double animationValue, Widget widget) {
-    Offset s = _getOffsetValue(values, animationValue, i);
-    return new Transform.translate(
+    final Offset s = _getOffsetValue(values!, animationValue, i);
+    return Transform.translate(
       offset: s,
       child: widget,
     );
@@ -340,29 +341,28 @@ class TranslateTransformBuilder extends TransformBuilder<Offset> {
 class CustomLayoutOption {
   final List<TransformBuilder> builders = [];
   final int startIndex;
-  final int stateCount;
+  final int? stateCount;
 
-  CustomLayoutOption({this.stateCount, this.startIndex})
+  CustomLayoutOption({this.stateCount, required this.startIndex})
       : assert(startIndex != null, stateCount != null);
 
   CustomLayoutOption addOpacity(List<double> values) {
-    builders.add(new OpacityTransformBuilder(values: values));
+    builders.add(OpacityTransformBuilder(values: values));
     return this;
   }
 
   CustomLayoutOption addTranslate(List<Offset> values) {
-    builders.add(new TranslateTransformBuilder(values: values));
+    builders.add(TranslateTransformBuilder(values: values));
     return this;
   }
 
   CustomLayoutOption addScale(List<double> values, Alignment alignment) {
-    builders
-        .add(new ScaleTransformBuilder(values: values, alignment: alignment));
+    builders.add(ScaleTransformBuilder(values: values, alignment: alignment));
     return this;
   }
 
   CustomLayoutOption addRotate(List<double> values) {
-    builders.add(new RotateTransformBuilder(values: values));
+    builders.add(RotateTransformBuilder(values: values));
     return this;
   }
 }
@@ -371,19 +371,19 @@ class _CustomLayoutSwiper extends _SubSwiper {
   final CustomLayoutOption option;
 
   _CustomLayoutSwiper(
-      {this.option,
-      double itemWidth,
-      bool loop,
-      double itemHeight,
-      ValueChanged<int> onIndexChanged,
-      Key key,
-      IndexedWidgetBuilder itemBuilder,
-      Curve curve,
-      int duration,
-      int index,
-      int itemCount,
-      Axis scrollDirection,
-      SwiperController controller})
+      {required this.option,
+      double? itemWidth,
+      bool? loop,
+      double? itemHeight,
+      ValueChanged<int>? onIndexChanged,
+      Key? key,
+      IndexedWidgetBuilder? itemBuilder,
+      Curve? curve,
+      int? duration,
+      int? index,
+      int? itemCount,
+      Axis? scrollDirection,
+      SwiperController? controller})
       : assert(option != null),
         super(
             loop: loop,
@@ -401,7 +401,7 @@ class _CustomLayoutSwiper extends _SubSwiper {
 
   @override
   State<StatefulWidget> createState() {
-    return new _CustomLayoutState();
+    return _CustomLayoutState();
   }
 }
 
@@ -411,12 +411,14 @@ class _CustomLayoutState extends _CustomLayoutStateBase<_CustomLayoutSwiper> {
     super.didChangeDependencies();
     _startIndex = widget.option.startIndex;
     _animationCount = widget.option.stateCount;
+    _currentIndex = widget.index ?? 0;
   }
 
   @override
   void didUpdateWidget(_CustomLayoutSwiper oldWidget) {
     _startIndex = widget.option.startIndex;
     _animationCount = widget.option.stateCount;
+    _currentIndex = widget.index ?? 0;
     super.didUpdateWidget(oldWidget);
   }
 
@@ -424,10 +426,10 @@ class _CustomLayoutState extends _CustomLayoutStateBase<_CustomLayoutSwiper> {
   Widget _buildItem(int index, int realIndex, double animationValue) {
     List<TransformBuilder> builders = widget.option.builders;
 
-    Widget child = new SizedBox(
+    Widget child = SizedBox(
         width: widget.itemWidth ?? double.infinity,
         height: widget.itemHeight ?? double.infinity,
-        child: widget.itemBuilder(context, realIndex));
+        child: widget.itemBuilder!(context, realIndex));
 
     for (int i = builders.length - 1; i >= 0; --i) {
       TransformBuilder builder = builders[i];
