@@ -10,7 +10,6 @@ import 'dart:async';
 import 'package:flutter_card_swipper/widgets/transformer_page_view/index_controller.dart';
 import 'package:flutter_card_swipper/widgets/transformer_page_view/transformer_page_view.dart';
 
-
 part 'custom_layout.dart';
 
 typedef void SwiperOnTap(int index);
@@ -33,27 +32,28 @@ class Swiper extends StatefulWidget {
   final bool outer;
 
   /// Inner item height, this property is valid if layout=STACK or layout=TINDER or LAYOUT=CUSTOM,
-  final double itemHeight;
+  final double? itemHeight;
 
   /// Inner item width, this property is valid if layout=STACK or layout=TINDER or LAYOUT=CUSTOM,
-  final double itemWidth;
+  final double? itemWidth;
 
   // height of the inside container,this property is valid when outer=true,otherwise the inside container size is controlled by parent widget
-  final double containerHeight;
+  final double? containerHeight;
+
   // width of the inside container,this property is valid when outer=true,otherwise the inside container size is controlled by parent widget
-  final double containerWidth;
+  final double? containerWidth;
 
   /// Build item on index
-  final IndexedWidgetBuilder itemBuilder;
+  final IndexedWidgetBuilder? itemBuilder;
 
   /// Support transform like Android PageView did
   /// `itemBuilder` and `transformItemBuilder` must have one not null
-  final PageTransformer transformer;
+  final PageTransformer? transformer;
 
   /// count of the display items
   final int itemCount;
 
-  final ValueChanged<int> onIndexChanged;
+  final ValueChanged<int>? onIndexChanged;
 
   ///auto play config
   final bool autoplay;
@@ -82,21 +82,21 @@ class Swiper extends StatefulWidget {
   final int index;
 
   ///Called when tap
-  final SwiperOnTap onTap;
+  final SwiperOnTap? onTap;
 
   ///The swiper pagination plugin
-  final SwiperPlugin pagination;
+  final SwiperPlugin? pagination;
 
   ///the swiper control button plugin
-  final SwiperPlugin control;
+  final SwiperPlugin? control;
 
   ///other plugins, you can custom your own plugin
-  final List<SwiperPlugin> plugins;
+  final List<SwiperPlugin>? plugins;
 
   ///
-  final SwiperController controller;
+  final SwiperController? controller;
 
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   ///
   final double viewportFraction;
@@ -105,13 +105,13 @@ class Swiper extends StatefulWidget {
   final SwiperLayout layout;
 
   /// this value is valid when layout == SwiperLayout.CUSTOM
-  final CustomLayoutOption customLayoutOption;
+  final CustomLayoutOption? customLayoutOption;
 
   // This value is valid when viewportFraction is set and < 1.0
-  final double scale;
+  final double? scale;
 
   // This value is valid when viewportFraction is set and < 1.0
-  final double fade;
+  final double? fade;
 
   final PageIndicatorLayout indicatorLayout;
 
@@ -121,14 +121,14 @@ class Swiper extends StatefulWidget {
 
     ///
     this.transformer,
-    @required this.itemCount,
+    required this.itemCount,
     this.autoplay: false,
     this.layout: SwiperLayout.DEFAULT,
     this.autoplayDelay: kDefaultAutoplayDelayMs,
     this.autoplayDisableOnInteraction: true,
     this.duration: kDefaultAutoplayTransactionDuration,
     this.onIndexChanged,
-    this.index,
+    this.index = 0,
     this.onTap,
     this.control,
     this.loop: true,
@@ -137,7 +137,7 @@ class Swiper extends StatefulWidget {
     this.pagination,
     this.plugins,
     this.physics,
-    Key key,
+    Key? key,
     this.controller,
     this.customLayoutOption,
 
@@ -151,7 +151,7 @@ class Swiper extends StatefulWidget {
     this.scale,
     this.fade,
   })  : assert(itemBuilder != null || transformer != null,
-            "itemBuilder and transformItemBuilder must not be both null"),
+            'itemBuilder and transformItemBuilder must not be both null'),
         assert(
             !loop ||
                 ((loop &&
@@ -160,41 +160,39 @@ class Swiper extends StatefulWidget {
                             indicatorLayout == PageIndicatorLayout.COLOR ||
                             indicatorLayout == PageIndicatorLayout.NONE)) ||
                     (loop && layout != SwiperLayout.DEFAULT)),
-            "Only support `PageIndicatorLayout.SCALE` and `PageIndicatorLayout.COLOR`when layout==SwiperLayout.DEFAULT in loop mode"),
+            'Only support `PageIndicatorLayout.SCALE` and `PageIndicatorLayout.COLOR`when layout==SwiperLayout.DEFAULT in loop mode'),
         super(key: key);
 
   factory Swiper.children({
-    List<Widget> children,
+    required List<Widget> children,
     bool autoplay: false,
-    PageTransformer transformer,
+    PageTransformer? transformer,
     int autoplayDelay: kDefaultAutoplayDelayMs,
     bool reverse: false,
     bool autoplayDisableOnInteraction: true,
     int duration: kDefaultAutoplayTransactionDuration,
-    ValueChanged<int> onIndexChanged,
-    int index,
-    SwiperOnTap onTap,
+    ValueChanged<int>? onIndexChanged,
+    int index = 0,
+    SwiperOnTap? onTap,
     bool loop: true,
     Curve curve: Curves.ease,
     Axis scrollDirection: Axis.horizontal,
-    SwiperPlugin pagination,
-    SwiperPlugin control,
-    List<SwiperPlugin> plugins,
-    SwiperController controller,
-    Key key,
-    CustomLayoutOption customLayoutOption,
-    ScrollPhysics physics,
-    double containerHeight,
-    double containerWidth,
+    SwiperPlugin? pagination,
+    SwiperPlugin? control,
+    List<SwiperPlugin>? plugins,
+    SwiperController? controller,
+    Key? key,
+    CustomLayoutOption? customLayoutOption,
+    ScrollPhysics? physics,
+    double? containerHeight,
+    double? containerWidth,
     double viewportFraction: 1.0,
-    double itemHeight,
-    double itemWidth,
+    double? itemHeight,
+    double? itemWidth,
     bool outer: false,
     double scale: 1.0,
   }) {
-    assert(children != null, "children must not be null");
-
-    return new Swiper(
+    return Swiper(
         transformer: transformer,
         customLayoutOption: customLayoutOption,
         containerHeight: containerHeight,
@@ -227,36 +225,36 @@ class Swiper extends StatefulWidget {
   }
 
   factory Swiper.list({
-    PageTransformer transformer,
-    List list,
-    CustomLayoutOption customLayoutOption,
-    SwiperDataBuilder builder,
+    PageTransformer? transformer,
+    required List list,
+    CustomLayoutOption? customLayoutOption,
+    SwiperDataBuilder? builder,
     bool autoplay: false,
     int autoplayDelay: kDefaultAutoplayDelayMs,
     bool reverse: false,
     bool autoplayDisableOnInteraction: true,
     int duration: kDefaultAutoplayTransactionDuration,
-    ValueChanged<int> onIndexChanged,
-    int index,
-    SwiperOnTap onTap,
+    ValueChanged<int>? onIndexChanged,
+    required int index,
+    SwiperOnTap? onTap,
     bool loop: true,
     Curve curve: Curves.ease,
     Axis scrollDirection: Axis.horizontal,
-    SwiperPlugin pagination,
-    SwiperPlugin control,
-    List<SwiperPlugin> plugins,
-    SwiperController controller,
-    Key key,
-    ScrollPhysics physics,
-    double containerHeight,
-    double containerWidth,
+    SwiperPlugin? pagination,
+    SwiperPlugin? control,
+    List<SwiperPlugin>? plugins,
+    SwiperController? controller,
+    Key? key,
+    ScrollPhysics? physics,
+    double? containerHeight,
+    double? containerWidth,
     double viewportFraction: 1.0,
-    double itemHeight,
-    double itemWidth,
+    double? itemHeight,
+    double? itemWidth,
     bool outer: false,
     double scale: 1.0,
   }) {
-    return new Swiper(
+    return Swiper(
         transformer: transformer,
         customLayoutOption: customLayoutOption,
         containerHeight: containerHeight,
@@ -283,27 +281,28 @@ class Swiper extends StatefulWidget {
         plugins: plugins,
         physics: physics,
         itemBuilder: (BuildContext context, int index) {
-          return builder(context, list[index], index);
+          return builder!(context, list[index], index);
         },
         itemCount: list.length);
   }
 
   @override
   State<StatefulWidget> createState() {
-    return new _SwiperState();
+    return _SwiperState();
   }
 }
 
 abstract class _SwiperTimerMixin extends State<Swiper> {
-  Timer _timer;
+  Timer? _timer;
 
-  SwiperController _controller;
+  late SwiperController _controller;
 
   @override
   void initState() {
-    _controller = widget.controller;
-    if (_controller == null) {
-      _controller = new SwiperController();
+    if (widget.controller == null) {
+      _controller = SwiperController();
+    } else {
+      _controller = widget.controller!;
     }
     _controller.addListener(_onController);
     _handleAutoplay();
@@ -333,8 +332,8 @@ abstract class _SwiperTimerMixin extends State<Swiper> {
   void didUpdateWidget(Swiper oldWidget) {
     if (_controller != oldWidget.controller) {
       if (oldWidget.controller != null) {
-        oldWidget.controller.removeListener(_onController);
-        _controller = oldWidget.controller;
+        oldWidget.controller!.removeListener(_onController);
+        _controller = oldWidget.controller!;
         _controller.addListener(_onController);
       }
     }
@@ -366,7 +365,7 @@ abstract class _SwiperTimerMixin extends State<Swiper> {
   }
 
   void _startAutoplay() {
-    assert(_timer == null, "Timer must be stopped before start!");
+    assert(_timer == null, 'Timer must be stopped before start!');
     _timer =
         Timer.periodic(Duration(milliseconds: widget.autoplayDelay), _onTimer);
   }
@@ -377,37 +376,37 @@ abstract class _SwiperTimerMixin extends State<Swiper> {
 
   void _stopAutoplay() {
     if (_timer != null) {
-      _timer.cancel();
+      _timer!.cancel();
       _timer = null;
     }
   }
 }
 
 class _SwiperState extends _SwiperTimerMixin {
-  int _activeIndex;
+  late int _activeIndex;
 
-  TransformerPageController _pageController;
+  TransformerPageController? _pageController;
 
   Widget _wrapTap(BuildContext context, int index) {
-    return new GestureDetector(
+    return GestureDetector(
       behavior: HitTestBehavior.opaque,
       onTap: () {
-        this.widget.onTap(index);
+        this.widget.onTap!(index);
       },
-      child: widget.itemBuilder(context, index),
+      child: widget.itemBuilder!(context, index),
     );
   }
 
   @override
   void initState() {
-    _activeIndex = widget.index ?? 0;
+    _activeIndex = widget.index;
     if (_isPageViewLayout()) {
-      _pageController = new TransformerPageController(
+      _pageController = TransformerPageController(
           initialPage: widget.index,
           loop: widget.loop,
           itemCount: widget.itemCount,
           reverse:
-              widget.transformer == null ? false : widget.transformer.reverse,
+              widget.transformer == null ? false : widget.transformer!.reverse,
           viewportFraction: widget.viewportFraction);
     }
     super.initState();
@@ -423,7 +422,7 @@ class _SwiperState extends _SwiperTimerMixin {
   }
 
   bool _getReverse(Swiper widget) =>
-      widget.transformer == null ? false : widget.transformer.reverse;
+      widget.transformer == null ? false : widget.transformer!.reverse;
 
   @override
   void didUpdateWidget(Swiper oldWidget) {
@@ -435,7 +434,7 @@ class _SwiperState extends _SwiperTimerMixin {
               widget.itemCount != oldWidget.itemCount ||
               widget.viewportFraction != oldWidget.viewportFraction ||
               _getReverse(widget) != _getReverse(oldWidget))) {
-        _pageController = new TransformerPageController(
+        _pageController = TransformerPageController(
             initialPage: widget.index,
             loop: widget.loop,
             itemCount: widget.itemCount,
@@ -446,7 +445,7 @@ class _SwiperState extends _SwiperTimerMixin {
       scheduleMicrotask(() {
         // So that we have a chance to do `removeListener` in child widgets.
         if (_pageController != null) {
-          _pageController.dispose();
+          _pageController!.dispose();
           _pageController = null;
         }
       });
@@ -461,12 +460,12 @@ class _SwiperState extends _SwiperTimerMixin {
       _activeIndex = index;
     });
     if (widget.onIndexChanged != null) {
-      widget.onIndexChanged(index);
+      widget.onIndexChanged!(index);
     }
   }
 
   Widget _buildSwiper() {
-    IndexedWidgetBuilder itemBuilder;
+    IndexedWidgetBuilder? itemBuilder;
     if (widget.onTap != null) {
       itemBuilder = _wrapTap;
     } else {
@@ -474,7 +473,7 @@ class _SwiperState extends _SwiperTimerMixin {
     }
 
     if (widget.layout == SwiperLayout.STACK) {
-      return new _StackSwiper(
+      return _StackSwiper(
         loop: widget.loop,
         itemWidth: widget.itemWidth,
         itemHeight: widget.itemHeight,
@@ -488,13 +487,13 @@ class _SwiperState extends _SwiperTimerMixin {
         scrollDirection: widget.scrollDirection,
       );
     } else if (_isPageViewLayout()) {
-      PageTransformer transformer = widget.transformer;
+      PageTransformer? transformer = widget.transformer;
       if (widget.scale != null || widget.fade != null) {
         transformer =
-            new ScaleAndFadeTransformer(scale: widget.scale, fade: widget.fade);
+            ScaleAndFadeTransformer(scale: widget.scale, fade: widget.fade);
       }
 
-      Widget child = new TransformerPageView(
+      final Widget child = TransformerPageView(
         pageController: _pageController,
         loop: widget.loop,
         itemCount: widget.itemCount,
@@ -502,7 +501,7 @@ class _SwiperState extends _SwiperTimerMixin {
         transformer: transformer,
         viewportFraction: widget.viewportFraction,
         index: _activeIndex,
-        duration: new Duration(milliseconds: widget.duration),
+        duration: Duration(milliseconds: widget.duration),
         scrollDirection: widget.scrollDirection,
         onPageChanged: _onIndexChanged,
         curve: widget.curve,
@@ -510,7 +509,7 @@ class _SwiperState extends _SwiperTimerMixin {
         controller: _controller,
       );
       if (widget.autoplayDisableOnInteraction && widget.autoplay) {
-        return new NotificationListener(
+        return NotificationListener(
           child: child,
           onNotification: (ScrollNotification notification) {
             if (notification is ScrollStartNotification) {
@@ -529,7 +528,7 @@ class _SwiperState extends _SwiperTimerMixin {
 
       return child;
     } else if (widget.layout == SwiperLayout.TINDER) {
-      return new _TinderSwiper(
+      return _TinderSwiper(
         loop: widget.loop,
         itemWidth: widget.itemWidth,
         itemHeight: widget.itemHeight,
@@ -543,9 +542,9 @@ class _SwiperState extends _SwiperTimerMixin {
         scrollDirection: widget.scrollDirection,
       );
     } else if (widget.layout == SwiperLayout.CUSTOM) {
-      return new _CustomLayoutSwiper(
+      return _CustomLayoutSwiper(
         loop: widget.loop,
-        option: widget.customLayoutOption,
+        option: widget.customLayoutOption!,
         itemWidth: widget.itemWidth,
         itemHeight: widget.itemHeight,
         itemCount: widget.itemCount,
@@ -558,13 +557,13 @@ class _SwiperState extends _SwiperTimerMixin {
         scrollDirection: widget.scrollDirection,
       );
     } else {
-      return new Container();
+      return Container();
     }
   }
 
-  SwiperPluginConfig _ensureConfig(SwiperPluginConfig config) {
+  SwiperPluginConfig? _ensureConfig(SwiperPluginConfig? config) {
     if (config == null) {
-      config = new SwiperPluginConfig(
+      config = SwiperPluginConfig(
           outer: widget.outer,
           itemCount: widget.itemCount,
           layout: widget.layout,
@@ -578,8 +577,8 @@ class _SwiperState extends _SwiperTimerMixin {
     return config;
   }
 
-  List<Widget> _ensureListForStack(
-      Widget swiper, List<Widget> listForStack, Widget widget) {
+  List<Widget>? _ensureListForStack(
+      Widget swiper, List<Widget>? listForStack, Widget widget) {
     if (listForStack == null) {
       listForStack = [swiper, widget];
     } else {
@@ -590,19 +589,19 @@ class _SwiperState extends _SwiperTimerMixin {
 
   @override
   Widget build(BuildContext context) {
-    Widget swiper = _buildSwiper();
-    List<Widget> listForStack;
-    SwiperPluginConfig config;
+    final Widget swiper = _buildSwiper();
+    List<Widget>? listForStack;
+    SwiperPluginConfig? config;
     if (widget.control != null) {
       //Stack
       config = _ensureConfig(config);
       listForStack = _ensureListForStack(
-          swiper, listForStack, widget.control.build(context, config));
+          swiper, listForStack, widget.control!.build(context, config));
     }
 
     if (widget.plugins != null) {
       config = _ensureConfig(config);
-      for (SwiperPlugin plugin in widget.plugins) {
+      for (SwiperPlugin plugin in widget.plugins!) {
         listForStack = _ensureListForStack(
             swiper, listForStack, plugin.build(context, config));
       }
@@ -611,17 +610,17 @@ class _SwiperState extends _SwiperTimerMixin {
       config = _ensureConfig(config);
       if (widget.outer) {
         return _buildOuterPagination(
-            widget.pagination,
-            listForStack == null ? swiper : new Stack(children: listForStack),
-            config);
+            widget.pagination as SwiperPagination,
+            listForStack == null ? swiper : Stack(children: listForStack),
+            config!);
       } else {
         listForStack = _ensureListForStack(
-            swiper, listForStack, widget.pagination.build(context, config));
+            swiper, listForStack, widget.pagination!.build(context, config));
       }
     }
 
     if (listForStack != null) {
-      return new Stack(
+      return Stack(
         children: listForStack,
       );
     }
@@ -636,15 +635,15 @@ class _SwiperState extends _SwiperTimerMixin {
     if (widget.containerHeight != null || widget.containerWidth != null) {
       list.add(swiper);
     } else {
-      list.add(new Expanded(child: swiper));
+      list.add(Expanded(child: swiper));
     }
 
-    list.add(new Align(
+    list.add(Align(
       alignment: Alignment.center,
       child: pagination.build(context, config),
     ));
 
-    return new Column(
+    return Column(
       children: list,
       crossAxisAlignment: CrossAxisAlignment.stretch,
       mainAxisSize: MainAxisSize.min,
@@ -653,20 +652,20 @@ class _SwiperState extends _SwiperTimerMixin {
 }
 
 abstract class _SubSwiper extends StatefulWidget {
-  final IndexedWidgetBuilder itemBuilder;
-  final int itemCount;
-  final int index;
-  final ValueChanged<int> onIndexChanged;
-  final SwiperController controller;
-  final int duration;
-  final Curve curve;
-  final double itemWidth;
-  final double itemHeight;
-  final bool loop;
-  final Axis scrollDirection;
+  final IndexedWidgetBuilder? itemBuilder;
+  final int? itemCount;
+  final int? index;
+  final ValueChanged<int>? onIndexChanged;
+  final SwiperController? controller;
+  final int? duration;
+  final Curve? curve;
+  final double? itemWidth;
+  final double? itemHeight;
+  final bool? loop;
+  final Axis? scrollDirection;
 
   _SubSwiper(
-      {Key key,
+      {Key? key,
       this.loop,
       this.itemHeight,
       this.itemWidth,
@@ -685,9 +684,9 @@ abstract class _SubSwiper extends StatefulWidget {
 
   int getCorrectIndex(int indexNeedsFix) {
     if (itemCount == 0) return 0;
-    int value = indexNeedsFix % itemCount;
+    int value = indexNeedsFix % itemCount!;
     if (value < 0) {
-      value += itemCount;
+      value += itemCount!;
     }
     return value;
   }
@@ -695,18 +694,18 @@ abstract class _SubSwiper extends StatefulWidget {
 
 class _TinderSwiper extends _SubSwiper {
   _TinderSwiper({
-    Key key,
-    Curve curve,
-    int duration,
-    SwiperController controller,
-    ValueChanged<int> onIndexChanged,
-    double itemHeight,
-    double itemWidth,
-    IndexedWidgetBuilder itemBuilder,
-    int index,
-    bool loop,
-    int itemCount,
-    Axis scrollDirection,
+    Key? key,
+    Curve? curve,
+    int? duration,
+    SwiperController? controller,
+    ValueChanged<int>? onIndexChanged,
+    double? itemHeight,
+    double? itemWidth,
+    IndexedWidgetBuilder? itemBuilder,
+    int? index,
+    bool? loop,
+    int? itemCount,
+    Axis? scrollDirection,
   })  : assert(itemWidth != null && itemHeight != null),
         super(
             loop: loop,
@@ -724,24 +723,24 @@ class _TinderSwiper extends _SubSwiper {
 
   @override
   State<StatefulWidget> createState() {
-    return new _TinderState();
+    return _TinderState();
   }
 }
 
 class _StackSwiper extends _SubSwiper {
   _StackSwiper({
-    Key key,
-    Curve curve,
-    int duration,
-    SwiperController controller,
-    ValueChanged<int> onIndexChanged,
-    double itemHeight,
-    double itemWidth,
-    IndexedWidgetBuilder itemBuilder,
-    int index,
-    bool loop,
-    int itemCount,
-    Axis scrollDirection,
+    Key? key,
+    Curve? curve,
+    int? duration,
+    SwiperController? controller,
+    ValueChanged<int>? onIndexChanged,
+    double? itemHeight,
+    double? itemWidth,
+    IndexedWidgetBuilder? itemBuilder,
+    int? index,
+    bool? loop,
+    int? itemCount,
+    Axis? scrollDirection,
   }) : super(
             loop: loop,
             key: key,
@@ -758,19 +757,19 @@ class _StackSwiper extends _SubSwiper {
 
   @override
   State<StatefulWidget> createState() {
-    return new _StackViewState();
+    return _StackViewState();
   }
 }
 
 class _TinderState extends _CustomLayoutStateBase<_TinderSwiper> {
-  List<double> scales;
-  List<double> offsetsX;
-  List<double> offsetsY;
-  List<double> opacity;
-  List<double> rotates;
+  late List<double> scales;
+  late List<double?> offsetsX;
+  late List<double?> offsetsY;
+  late List<double> opacity;
+  late List<double> rotates;
 
   double getOffsetY(double scale) {
-    return widget.itemHeight - widget.itemHeight * scale;
+    return widget.itemHeight! - widget.itemHeight! * scale;
   }
 
   @override
@@ -823,30 +822,30 @@ class _TinderState extends _CustomLayoutStateBase<_TinderSwiper> {
 
   @override
   Widget _buildItem(int i, int realIndex, double animationValue) {
-    double s = _getValue(scales, animationValue, i);
-    double f = _getValue(offsetsX, animationValue, i);
-    double fy = _getValue(offsetsY, animationValue, i);
-    double o = _getValue(opacity, animationValue, i);
-    double a = _getValue(rotates, animationValue, i);
+    final double s = _getValue(scales, animationValue, i)!;
+    final double f = _getValue(offsetsX, animationValue, i)!;
+    final double fy = _getValue(offsetsY, animationValue, i)!;
+    final double o = _getValue(opacity, animationValue, i)!;
+    final double a = _getValue(rotates, animationValue, i)!;
 
-    Alignment alignment = widget.scrollDirection == Axis.horizontal
+    final Alignment alignment = widget.scrollDirection == Axis.horizontal
         ? Alignment.bottomCenter
         : Alignment.centerLeft;
 
-    return new Opacity(
+    return Opacity(
       opacity: o,
-      child: new Transform.rotate(
+      child: Transform.rotate(
         angle: a / 180.0,
-        child: new Transform.translate(
-          key: new ValueKey<int>(_currentIndex + i),
-          offset: new Offset(f, fy),
-          child: new Transform.scale(
+        child: Transform.translate(
+          key: ValueKey<int>(_currentIndex + i),
+          offset: Offset(f, fy),
+          child: Transform.scale(
             scale: s,
             alignment: alignment,
-            child: new SizedBox(
+            child: SizedBox(
               width: widget.itemWidth ?? double.infinity,
               height: widget.itemHeight ?? double.infinity,
-              child: widget.itemBuilder(context, realIndex),
+              child: widget.itemBuilder!(context, realIndex),
             ),
           ),
         ),
@@ -856,9 +855,10 @@ class _TinderState extends _CustomLayoutStateBase<_TinderSwiper> {
 }
 
 class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
-  List<double> scales;
-  List<double> offsets;
-  List<double> opacity;
+  late List<double> scales;
+  late List<double?> offsets;
+  late List<double> opacity;
+
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
@@ -866,10 +866,10 @@ class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
 
   void _updateValues() {
     if (widget.scrollDirection == Axis.horizontal) {
-      double space = (_swiperWidth - widget.itemWidth) / 2;
+      double space = (_swiperWidth! - widget.itemWidth!) / 2;
       offsets = [-space, -space / 3 * 2, -space / 3, 0.0, _swiperWidth];
     } else {
-      double space = (_swiperHeight - widget.itemHeight) / 2;
+      final double space = (_swiperHeight! - widget.itemHeight!) / 2;
       offsets = [-space, -space / 3 * 2, -space / 3, 0.0, _swiperHeight];
     }
   }
@@ -897,30 +897,30 @@ class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
 
   @override
   Widget _buildItem(int i, int realIndex, double animationValue) {
-    double s = _getValue(scales, animationValue, i);
-    double f = _getValue(offsets, animationValue, i);
-    double o = _getValue(opacity, animationValue, i);
+    final double s = _getValue(scales, animationValue, i)!;
+    final double? f = _getValue(offsets, animationValue, i);
+    final double o = _getValue(opacity, animationValue, i)!;
 
-    Offset offset = widget.scrollDirection == Axis.horizontal
-        ? new Offset(f, 0.0)
-        : new Offset(0.0, f);
+    final offset = widget.scrollDirection == Axis.horizontal
+        ? Offset(f!, 0.0)
+        : Offset(0.0, f!);
 
-    Alignment alignment = widget.scrollDirection == Axis.horizontal
+    final alignment = widget.scrollDirection == Axis.horizontal
         ? Alignment.centerLeft
         : Alignment.topCenter;
 
-    return new Opacity(
+    return Opacity(
       opacity: o,
-      child: new Transform.translate(
-        key: new ValueKey<int>(_currentIndex + i),
+      child: Transform.translate(
+        key: ValueKey<int>(_currentIndex + i),
         offset: offset,
-        child: new Transform.scale(
+        child: Transform.scale(
           scale: s,
           alignment: alignment,
-          child: new SizedBox(
+          child: SizedBox(
             width: widget.itemWidth ?? double.infinity,
             height: widget.itemHeight ?? double.infinity,
-            child: widget.itemBuilder(context, realIndex),
+            child: widget.itemBuilder!(context, realIndex),
           ),
         ),
       ),
@@ -929,31 +929,31 @@ class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
 }
 
 class ScaleAndFadeTransformer extends PageTransformer {
-  final double _scale;
-  final double _fade;
+  final double? _scale;
+  final double? _fade;
 
-  ScaleAndFadeTransformer({double fade: 0.3, double scale: 0.8})
+  ScaleAndFadeTransformer({double? fade: 0.3, double? scale: 0.8})
       : _fade = fade,
         _scale = scale;
 
   @override
   Widget transform(Widget item, TransformInfo info) {
-    double position = info.position;
+    double? position = info.position;
     Widget child = item;
     if (_scale != null) {
-      double scaleFactor = (1 - position.abs()) * (1 - _scale);
-      double scale = _scale + scaleFactor;
+      double scaleFactor = (1 - position!.abs()) * (1 - _scale!);
+      double scale = _scale! + scaleFactor;
 
-      child = new Transform.scale(
+      child = Transform.scale(
         scale: scale,
         child: item,
       );
     }
 
     if (_fade != null) {
-      double fadeFactor = (1 - position.abs()) * (1 - _fade);
-      double opacity = _fade + fadeFactor;
-      child = new Opacity(
+      double fadeFactor = (1 - position!.abs()) * (1 - _fade!);
+      double opacity = _fade! + fadeFactor;
+      child = Opacity(
         opacity: opacity,
         child: child,
       );
