@@ -6,10 +6,10 @@ import 'package:flutter_card_swipper/widgets/flutter_page_indicator/flutter_page
 
 class FractionPaginationBuilder extends SwiperPlugin {
   ///color ,if set null , will be Theme.of(context).scaffoldBackgroundColor
-  final Color color;
+  final Color? color;
 
   ///color when active,if set null , will be Theme.of(context).primaryColor
-  final Color activeColor;
+  final Color? activeColor;
 
   ////font size
   final double fontSize;
@@ -17,7 +17,7 @@ class FractionPaginationBuilder extends SwiperPlugin {
   ///font size when active
   final double activeFontSize;
 
-  final Key key;
+  final Key? key;
 
   const FractionPaginationBuilder(
       {this.color,
@@ -27,41 +27,41 @@ class FractionPaginationBuilder extends SwiperPlugin {
       this.activeFontSize: 35.0});
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
     ThemeData themeData = Theme.of(context);
     Color activeColor = this.activeColor ?? themeData.primaryColor;
     Color color = this.color ?? themeData.scaffoldBackgroundColor;
 
-    if (Axis.vertical == config.scrollDirection) {
-      return new Column(
+    if (Axis.vertical == config!.scrollDirection) {
+      return Column(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          new Text(
-            "${config.activeIndex + 1}",
+          Text(
+            '${config.activeIndex! + 1}',
             style: TextStyle(color: activeColor, fontSize: activeFontSize),
           ),
-          new Text(
-            "/",
+          Text(
+            '/',
             style: TextStyle(color: color, fontSize: fontSize),
           ),
-          new Text(
-            "${config.itemCount}",
+          Text(
+            '${config.itemCount}',
             style: TextStyle(color: color, fontSize: fontSize),
           )
         ],
       );
     } else {
-      return new Row(
+      return Row(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          new Text(
-            "${config.activeIndex + 1}",
+          Text(
+            '${config.activeIndex! + 1}',
             style: TextStyle(color: activeColor, fontSize: activeFontSize),
           ),
-          new Text(
-            " / ${config.itemCount}",
+          Text(
+            ' / ${config.itemCount}',
             style: TextStyle(color: color, fontSize: fontSize),
           )
         ],
@@ -72,10 +72,10 @@ class FractionPaginationBuilder extends SwiperPlugin {
 
 class RectSwiperPaginationBuilder extends SwiperPlugin {
   ///color when current index,if set null , will be Theme.of(context).primaryColor
-  final Color activeColor;
+  final Color? activeColor;
 
   ///,if set null , will be Theme.of(context).scaffoldBackgroundColor
-  final Color color;
+  final Color? color;
 
   ///Size of the rect when activate
   final Size activeSize;
@@ -86,7 +86,7 @@ class RectSwiperPaginationBuilder extends SwiperPlugin {
   /// Space between rects
   final double space;
 
-  final Key key;
+  final Key? key;
 
   const RectSwiperPaginationBuilder(
       {this.activeColor,
@@ -97,20 +97,20 @@ class RectSwiperPaginationBuilder extends SwiperPlugin {
       this.space: 3.0});
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
     ThemeData themeData = Theme.of(context);
     Color activeColor = this.activeColor ?? themeData.primaryColor;
     Color color = this.color ?? themeData.scaffoldBackgroundColor;
 
     List<Widget> list = [];
 
-    if (config.itemCount > 20) {
+    if (config!.itemCount! > 20) {
       print(
-          "The itemCount is too big, we suggest use FractionPaginationBuilder instead of DotSwiperPaginationBuilder in this sitituation");
+          'The itemCount is too big, we suggest use FractionPaginationBuilder instead of DotSwiperPaginationBuilder in this sitituation');
     }
 
-    int itemCount = config.itemCount;
-    int activeIndex = config.activeIndex;
+    int itemCount = config.itemCount!;
+    int? activeIndex = config.activeIndex;
 
     for (int i = 0; i < itemCount; ++i) {
       bool active = i == activeIndex;
@@ -120,20 +120,20 @@ class RectSwiperPaginationBuilder extends SwiperPlugin {
         height: size.height,
         child: Container(
           color: active ? activeColor : color,
-          key: Key("pagination_$i"),
+          key: Key('pagination_$i'),
           margin: EdgeInsets.all(space),
         ),
       ));
     }
 
     if (config.scrollDirection == Axis.vertical) {
-      return new Column(
+      return Column(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
       );
     } else {
-      return new Row(
+      return Row(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
@@ -144,10 +144,10 @@ class RectSwiperPaginationBuilder extends SwiperPlugin {
 
 class DotSwiperPaginationBuilder extends SwiperPlugin {
   ///color when current index,if set null , will be Theme.of(context).primaryColor
-  final Color activeColor;
+  final Color? activeColor;
 
   ///,if set null , will be Theme.of(context).scaffoldBackgroundColor
-  final Color color;
+  final Color? color;
 
   ///Size of the dot when activate
   final double activeSize;
@@ -158,7 +158,7 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
   /// Space between dots
   final double space;
 
-  final Key key;
+  final Key? key;
 
   const DotSwiperPaginationBuilder(
       {this.activeColor,
@@ -169,13 +169,13 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
       this.space: 3.0});
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
-    if (config.itemCount > 20) {
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
+    if (config!.itemCount! > 20) {
       print(
-          "The itemCount is too big, we suggest use FractionPaginationBuilder instead of DotSwiperPaginationBuilder in this sitituation");
+          'The itemCount is too big, we suggest use FractionPaginationBuilder instead of DotSwiperPaginationBuilder in this sitituation');
     }
-    Color activeColor = this.activeColor;
-    Color color = this.color;
+    Color? activeColor = this.activeColor;
+    Color? color = this.color;
 
     if (activeColor == null || color == null) {
       ThemeData themeData = Theme.of(context);
@@ -185,7 +185,7 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
 
     if (config.indicatorLayout != PageIndicatorLayout.NONE &&
         config.layout == SwiperLayout.DEFAULT) {
-      return new PageIndicator(
+      return PageIndicator(
         count: config.itemCount,
         controller: config.pageController,
         layout: config.indicatorLayout,
@@ -198,13 +198,13 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
 
     List<Widget> list = [];
 
-    int itemCount = config.itemCount;
-    int activeIndex = config.activeIndex;
+    int itemCount = config.itemCount!;
+    int? activeIndex = config.activeIndex;
 
     for (int i = 0; i < itemCount; ++i) {
       bool active = i == activeIndex;
       list.add(Container(
-        key: Key("pagination_$i"),
+        key: Key('pagination_$i'),
         margin: EdgeInsets.all(space),
         child: ClipOval(
           child: Container(
@@ -217,13 +217,13 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
     }
 
     if (config.scrollDirection == Axis.vertical) {
-      return new Column(
+      return Column(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
       );
     } else {
-      return new Row(
+      return Row(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
@@ -233,15 +233,15 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
 }
 
 typedef Widget SwiperPaginationBuilder(
-    BuildContext context, SwiperPluginConfig config);
+    BuildContext context, SwiperPluginConfig? config);
 
 class SwiperCustomPagination extends SwiperPlugin {
   final SwiperPaginationBuilder builder;
 
-  SwiperCustomPagination({@required this.builder});
+  SwiperCustomPagination({required this.builder});
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
     return builder(context, config);
   }
 }
@@ -257,7 +257,7 @@ class SwiperPagination extends SwiperPlugin {
 
   /// Alignment.bottomCenter by default when scrollDirection== Axis.horizontal
   /// Alignment.centerRight by default when scrollDirection== Axis.vertical
-  final Alignment alignment;
+  final Alignment? alignment;
 
   /// Distance between pagination and the container
   final EdgeInsetsGeometry margin;
@@ -265,7 +265,7 @@ class SwiperPagination extends SwiperPlugin {
   /// Build the widet
   final SwiperPlugin builder;
 
-  final Key key;
+  final Key? key;
 
   const SwiperPagination(
       {this.alignment,
@@ -273,17 +273,17 @@ class SwiperPagination extends SwiperPlugin {
       this.margin: const EdgeInsets.all(10.0),
       this.builder: SwiperPagination.dots});
 
-  Widget build(BuildContext context, SwiperPluginConfig config) {
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
     Alignment alignment = this.alignment ??
-        (config.scrollDirection == Axis.horizontal
+        (config!.scrollDirection == Axis.horizontal
             ? Alignment.bottomCenter
             : Alignment.centerRight);
     Widget child = Container(
       margin: margin,
       child: this.builder.build(context, config),
     );
-    if (!config.outer) {
-      child = new Align(
+    if (!config!.outer!) {
+      child = Align(
         key: key,
         alignment: alignment,
         child: child,
