@@ -17,15 +17,13 @@ class WarmPainter extends BasePainter {
 
       double left = index * distance + distance * (progress - 0.5) * 2;
       canvas.drawRRect(
-          new RRect.fromLTRBR(
-              left, 0.0, right, size, new Radius.circular(radius)),
+          RRect.fromLTRBR(left, 0.0, right, size, Radius.circular(radius)),
           _paint);
     } else {
       double right = start + size + distance * progress * 2;
 
       canvas.drawRRect(
-          new RRect.fromLTRBR(
-              start, 0.0, right, size, new Radius.circular(radius)),
+          RRect.fromLTRBR(start, 0.0, right, size, Radius.circular(radius)),
           _paint);
     }
   }
@@ -45,8 +43,8 @@ class DropPainter extends BasePainter {
     //lerp(begin, end, progress)
 
     canvas.drawCircle(
-        new Offset(radius + ((page) * (size + space)),
-            radius - dropHeight * (1 - rate)),
+        Offset(
+            radius + (page * (size + space)), radius - dropHeight * (1 - rate)),
         radius * (scale + rate * (1.0 - scale)),
         _paint);
   }
@@ -59,15 +57,15 @@ class NonePainter extends BasePainter {
   @override
   void draw(Canvas canvas, double space, double size, double radius) {
     double progress = page - index;
-    double secondOffset = index == widget.count - 1
+    double secondOffset = index == widget.count! - 1
         ? radius
         : radius + ((index + 1) * (size + space));
 
     if (progress > 0.5) {
-      canvas.drawCircle(new Offset(secondOffset, radius), radius, _paint);
+      canvas.drawCircle(Offset(secondOffset, radius), radius, _paint);
     } else {
-      canvas.drawCircle(new Offset(radius + (index * (size + space)), radius),
-          radius, _paint);
+      canvas.drawCircle(
+          Offset(radius + (index * (size + space)), radius), radius, _paint);
     }
   }
 }
@@ -79,7 +77,7 @@ class SlidePainter extends BasePainter {
   @override
   void draw(Canvas canvas, double space, double size, double radius) {
     canvas.drawCircle(
-        new Offset(radius + (page * (size + space)), radius), radius, _paint);
+        Offset(radius + (page * (size + space)), radius), radius, _paint);
   }
 }
 
@@ -89,11 +87,11 @@ class ScalePainter extends BasePainter {
 
   // 连续的两个点，含有最后一个和第一个
   @override
-  bool _shouldSkip(int i) {
-    if (index == widget.count - 1) {
+  bool _shouldSkip(int? i) {
+    if (index == widget.count! - 1) {
       return i == 0 || i == index;
     }
-    return (i == index || i == index + 1);
+    return i == index || i == index + 1;
   }
 
   @override
@@ -102,11 +100,11 @@ class ScalePainter extends BasePainter {
     double space = widget.space;
     double size = widget.size;
     double radius = size / 2;
-    for (int i = 0, c = widget.count; i < c; ++i) {
+    for (int? i = 0, c = widget.count; i! < c!; ++i) {
       if (_shouldSkip(i)) {
         continue;
       }
-      canvas.drawCircle(new Offset(i * (size + space) + radius, radius),
+      canvas.drawCircle(Offset(i * (size + space) + radius, radius),
           radius * widget.scale, _paint);
     }
 
@@ -116,18 +114,18 @@ class ScalePainter extends BasePainter {
 
   @override
   void draw(Canvas canvas, double space, double size, double radius) {
-    double secondOffset = index == widget.count - 1
+    double secondOffset = index == widget.count! - 1
         ? radius
         : radius + ((index + 1) * (size + space));
 
     double progress = page - index;
-    _paint.color = Color.lerp(widget.activeColor, widget.color, progress);
+    _paint.color = Color.lerp(widget.activeColor, widget.color, progress)!;
     //last
-    canvas.drawCircle(new Offset(radius + (index * (size + space)), radius),
+    canvas.drawCircle(Offset(radius + (index * (size + space)), radius),
         lerp(radius, radius * widget.scale, progress), _paint);
     //first
-    _paint.color = Color.lerp(widget.color, widget.activeColor, progress);
-    canvas.drawCircle(new Offset(secondOffset, radius),
+    _paint.color = Color.lerp(widget.color, widget.activeColor, progress)!;
+    canvas.drawCircle(Offset(secondOffset, radius),
         lerp(radius * widget.scale, radius, progress), _paint);
   }
 }
@@ -136,29 +134,28 @@ class ColorPainter extends BasePainter {
   ColorPainter(PageIndicator widget, double page, int index, Paint paint)
       : super(widget, page, index, paint);
 
-  // 连续的两个点，含有最后一个和第一个
   @override
-  bool _shouldSkip(int i) {
-    if (index == widget.count - 1) {
+  bool _shouldSkip(int? i) {
+    if (index == widget.count! - 1) {
       return i == 0 || i == index;
     }
-    return (i == index || i == index + 1);
+    return i == index || i == index + 1;
   }
 
   @override
   void draw(Canvas canvas, double space, double size, double radius) {
-    double progress = page - index;
-    double secondOffset = index == widget.count - 1
+    final double progress = page - index;
+    final double secondOffset = index == widget.count! - 1
         ? radius
         : radius + ((index + 1) * (size + space));
 
-    _paint.color = Color.lerp(widget.activeColor, widget.color, progress);
+    _paint.color = Color.lerp(widget.activeColor, widget.color, progress)!;
     //left
     canvas.drawCircle(
-        new Offset(radius + (index * (size + space)), radius), radius, _paint);
+        Offset(radius + (index * (size + space)), radius), radius, _paint);
     //right
-    _paint.color = Color.lerp(widget.color, widget.activeColor, progress);
-    canvas.drawCircle(new Offset(secondOffset, radius), radius, _paint);
+    _paint.color = Color.lerp(widget.color, widget.activeColor, progress)!;
+    canvas.drawCircle(Offset(secondOffset, radius), radius, _paint);
   }
 }
 
@@ -176,7 +173,7 @@ abstract class BasePainter extends CustomPainter {
 
   void draw(Canvas canvas, double space, double size, double radius);
 
-  bool _shouldSkip(int index) {
+  bool _shouldSkip(int? index) {
     return false;
   }
 
@@ -188,12 +185,12 @@ abstract class BasePainter extends CustomPainter {
     double space = widget.space;
     double size = widget.size;
     double radius = size / 2;
-    for (int i = 0, c = widget.count; i < c; ++i) {
+    for (int? i = 0, c = widget.count; i! < c!; ++i) {
       if (_shouldSkip(i)) {
         continue;
       }
       canvas.drawCircle(
-          new Offset(i * (size + space) + radius, radius), radius, _paint);
+          Offset(i * (size + space) + radius, radius), radius, _paint);
     }
 
     double page = this.page;
@@ -212,57 +209,59 @@ abstract class BasePainter extends CustomPainter {
 
 class _PageIndicatorState extends State<PageIndicator> {
   int index = 0;
-  Paint _paint = new Paint();
+  Paint _paint = Paint();
 
   BasePainter _createPainer() {
     switch (widget.layout) {
       case PageIndicatorLayout.NONE:
-        return new NonePainter(
-            widget, widget.controller.page ?? 0.0, index, _paint);
+        return NonePainter(
+            widget, widget.controller!.page ?? 0.0, index, _paint);
       case PageIndicatorLayout.SLIDE:
-        return new SlidePainter(
-            widget, widget.controller.page ?? 0.0, index, _paint);
+        return SlidePainter(
+            widget, widget.controller!.page ?? 0.0, index, _paint);
       case PageIndicatorLayout.WARM:
-        return new WarmPainter(
-            widget, widget.controller.page ?? 0.0, index, _paint);
+        return WarmPainter(
+            widget, widget.controller!.page ?? 0.0, index, _paint);
       case PageIndicatorLayout.COLOR:
-        return new ColorPainter(
-            widget, widget.controller.page ?? 0.0, index, _paint);
+        return ColorPainter(
+            widget, widget.controller!.page ?? 0.0, index, _paint);
       case PageIndicatorLayout.SCALE:
-        return new ScalePainter(
-            widget, widget.controller.page ?? 0.0, index, _paint);
+        return ScalePainter(
+            widget, widget.controller!.page ?? 0.0, index, _paint);
       case PageIndicatorLayout.DROP:
-        return new DropPainter(
-            widget, widget.controller.page ?? 0.0, index, _paint);
+        return DropPainter(
+            widget, widget.controller!.page ?? 0.0, index, _paint);
       default:
-        throw new Exception("Not a valid layout");
+        throw Exception('Not a valid layout');
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    Widget child = new SizedBox(
-      width: widget.count * widget.size + (widget.count - 1) * widget.space,
+    Widget child = SizedBox(
+      width: widget.count! > 0
+          ? widget.count! * widget.size + (widget.count! - 1) * widget.space
+          : 0.0,
       height: widget.size,
-      child: new CustomPaint(
+      child: CustomPaint(
         painter: _createPainer(),
       ),
     );
 
     if (widget.layout == PageIndicatorLayout.SCALE ||
         widget.layout == PageIndicatorLayout.COLOR) {
-      child = new ClipRect(
+      child = ClipRect(
         child: child,
       );
     }
 
-    return new IgnorePointer(
+    return IgnorePointer(
       child: child,
     );
   }
 
   void _onController() {
-    double page = widget.controller.page ?? 0.0;
+    double page = widget.controller!.page ?? 0.0;
     index = page.floor();
 
     setState(() {});
@@ -270,22 +269,22 @@ class _PageIndicatorState extends State<PageIndicator> {
 
   @override
   void initState() {
-    widget.controller.addListener(_onController);
+    widget.controller!.addListener(_onController);
     super.initState();
   }
 
   @override
   void didUpdateWidget(PageIndicator oldWidget) {
     if (widget.controller != oldWidget.controller) {
-      oldWidget.controller.removeListener(_onController);
-      widget.controller.addListener(_onController);
+      oldWidget.controller!.removeListener(_onController);
+      widget.controller!.addListener(_onController);
     }
     super.didUpdateWidget(oldWidget);
   }
 
   @override
   void dispose() {
-    widget.controller.removeListener(_onController);
+    widget.controller!.removeListener(_onController);
     super.dispose();
   }
 }
@@ -307,7 +306,7 @@ class PageIndicator extends StatefulWidget {
   final double space;
 
   /// count of dots
-  final int count;
+  final int? count;
 
   /// active color
   final Color activeColor;
@@ -316,7 +315,7 @@ class PageIndicator extends StatefulWidget {
   final Color color;
 
   /// layout of the dots,default is [PageIndicatorLayout.SLIDE]
-  final PageIndicatorLayout layout;
+  final PageIndicatorLayout? layout;
 
   // Only valid when layout==PageIndicatorLayout.scale
   final double scale;
@@ -324,12 +323,12 @@ class PageIndicator extends StatefulWidget {
   // Only valid when layout==PageIndicatorLayout.drop
   final double dropHeight;
 
-  final PageController controller;
+  final PageController? controller;
 
   final double activeSize;
 
   PageIndicator({
-    Key key,
+    Key? key,
     this.size: 20.0,
     this.space: 5.0,
     this.count,
@@ -344,6 +343,6 @@ class PageIndicator extends StatefulWidget {
 
   @override
   State<StatefulWidget> createState() {
-    return new _PageIndicatorState();
+    return _PageIndicatorState();
   }
 }
