@@ -8,33 +8,33 @@ class IndexController extends ChangeNotifier {
   static const int PREVIOUS = -1;
   static const int MOVE = 0;
 
-  Completer _completer;
+  late Completer _completer;
 
-  int index;
-  bool animation;
-  int event;
+  int? index;
+  bool? animation;
+  int? event;
 
   Future move(int index, {bool animation: true}) {
-    this.animation = animation ?? true;
+    this.animation = animation;
     this.index = index;
     this.event = MOVE;
-    _completer = new Completer();
+    _completer = Completer();
     notifyListeners();
     return _completer.future;
   }
 
   Future next({bool animation: true}) {
     this.event = NEXT;
-    this.animation = animation ?? true;
-    _completer = new Completer();
+    this.animation = animation;
+    _completer = Completer();
     notifyListeners();
     return _completer.future;
   }
 
   Future previous({bool animation: true}) {
     this.event = PREVIOUS;
-    this.animation = animation ?? true;
-    _completer = new Completer();
+    this.animation = animation;
+    _completer = Completer();
     notifyListeners();
     return _completer.future;
   }
