@@ -14,21 +14,20 @@ class ColorPainter extends CustomPainter {
   void paint(Canvas canvas, Size size) {
     int index = info.fromIndex;
     _paint.color = colors[index];
-    canvas.drawRect(
-        new Rect.fromLTWH(0.0, 0.0, size.width, size.height), _paint);
-    if (info.done) {
+    canvas.drawRect(Rect.fromLTWH(0.0, 0.0, size.width, size.height), _paint);
+    if (info.done!) {
       return;
     }
     int alpha;
     int color;
     double opacity;
-    double position = info.position;
-    if (info.forward) {
+    double? position = info.position;
+    if (info.forward!) {
       if (index < colors.length - 1) {
         color = colors[index + 1].value & 0x00ffffff;
-        opacity = (position <= 0
-            ? (-position / info.viewportFraction)
-            : 1 - position / info.viewportFraction);
+        opacity = (position! <= 0
+            ? (-position / info.viewportFraction!)
+            : 1 - position / info.viewportFraction!);
         if (opacity > 1) {
           opacity -= 1.0;
         }
@@ -37,16 +36,16 @@ class ColorPainter extends CustomPainter {
         }
         alpha = (0xff * opacity).toInt();
 
-        _paint.color = new Color((alpha << 24) | color);
+        _paint.color = Color((alpha << 24) | color);
         canvas.drawRect(
-            new Rect.fromLTWH(0.0, 0.0, size.width, size.height), _paint);
+            Rect.fromLTWH(0.0, 0.0, size.width, size.height), _paint);
       }
     } else {
       if (index > 0) {
         color = colors[index - 1].value & 0x00ffffff;
-        opacity = (position > 0
-            ? position / info.viewportFraction
-            : (1 + position / info.viewportFraction));
+        opacity = (position! > 0
+            ? position / info.viewportFraction!
+            : (1 + position / info.viewportFraction!));
         if (opacity > 1) {
           opacity -= 1.0;
         }
@@ -55,9 +54,9 @@ class ColorPainter extends CustomPainter {
         }
         alpha = (0xff * opacity).toInt();
 
-        _paint.color = new Color((alpha << 24) | color);
+        _paint.color = Color((alpha << 24) | color);
         canvas.drawRect(
-            new Rect.fromLTWH(0.0, 0.0, size.width, size.height), _paint);
+            Rect.fromLTWH(0.0, 0.0, size.width, size.height), _paint);
       }
     }
   }
@@ -69,12 +68,12 @@ class ColorPainter extends CustomPainter {
 }
 
 class _ParallaxColorState extends State<ParallaxColor> {
-  Paint paint = new Paint();
+  Paint paint = Paint();
 
   @override
   Widget build(BuildContext context) {
-    return new CustomPaint(
-      painter: new ColorPainter(paint, widget.info, widget.colors),
+    return CustomPaint(
+      painter: ColorPainter(paint, widget.info, widget.colors),
       child: widget.child,
     );
   }
@@ -88,14 +87,14 @@ class ParallaxColor extends StatefulWidget {
   final TransformInfo info;
 
   ParallaxColor({
-    @required this.colors,
-    @required this.info,
-    @required this.child,
+    required this.colors,
+    required this.info,
+    required this.child,
   });
 
   @override
   State<StatefulWidget> createState() {
-    return new _ParallaxColorState();
+    return _ParallaxColorState();
   }
 }
 
@@ -106,8 +105,8 @@ class ParallaxContainer extends StatelessWidget {
   final double opacityFactor;
 
   ParallaxContainer(
-      {@required this.child,
-      @required this.position,
+      {required this.child,
+      required this.position,
       this.translationFactor: 100.0,
       this.opacityFactor: 1.0})
       : assert(position != null),
@@ -117,8 +116,8 @@ class ParallaxContainer extends StatelessWidget {
   Widget build(BuildContext context) {
     return Opacity(
       opacity: (1 - position.abs()).clamp(0.0, 1.0) * opacityFactor,
-      child: new Transform.translate(
-        offset: new Offset(position * translationFactor, 0.0),
+      child: Transform.translate(
+        offset: Offset(position * translationFactor, 0.0),
         child: child,
       ),
     );
@@ -129,7 +128,8 @@ class ParallaxImage extends StatelessWidget {
   final Image image;
   final double imageFactor;
 
-  ParallaxImage.asset(String name, {double position, this.imageFactor: 0.3})
+  ParallaxImage.asset(String name,
+      {required double position, this.imageFactor: 0.3})
       : assert(imageFactor != null),
         image = Image.asset(name,
             fit: BoxFit.cover,
