@@ -9,6 +9,7 @@ import 'package:flutter_card_swipper/widgets/transformer_page_view/index_control
 ///
 ///
 ///
+
 const int kMaxValue = 2000000000;
 const int kMiddleValue = 1000000000;
 
@@ -17,10 +18,10 @@ const int kDefaultTransactionDuration = 300;
 
 class TransformInfo {
   /// The `width` of the `TransformerPageView`
-  final double width;
+  final double? width;
 
   /// The `height` of the `TransformerPageView`
-  final double height;
+  final double? height;
 
   /// The `position` of the widget pass to [PageTransformer.transform]
   ///  A `position` describes how visible the widget is.
@@ -29,10 +30,10 @@ class TransformInfo {
   ///  The widge in the right ,may be hidden, of the screen's position is greater than 0.0,  1.0 when out of the screen
   ///
   ///
-  final double position;
+  final double? position;
 
   /// The `index` of the widget pass to [PageTransformer.transform]
-  final int index;
+  final int? index;
 
   /// The `activeIndex` of the PageView
   final int activeIndex;
@@ -42,24 +43,24 @@ class TransformInfo {
   final int fromIndex;
 
   /// Next `index` is greater than this `index`
-  final bool forward;
+  final bool? forward;
 
   /// User drag is done.
-  final bool done;
+  final bool? done;
 
   /// Same as [TransformerPageView.viewportFraction]
-  final double viewportFraction;
+  final double? viewportFraction;
 
   /// Copy from [TransformerPageView.scrollDirection]
-  final Axis scrollDirection;
+  final Axis? scrollDirection;
 
   TransformInfo(
       {this.index,
       this.position,
       this.width,
       this.height,
-      this.activeIndex,
-      this.fromIndex,
+      required this.activeIndex,
+      required this.fromIndex,
       this.forward,
       this.done,
       this.viewportFraction,
@@ -81,7 +82,7 @@ typedef Widget PageTransformerBuilderCallback(Widget child, TransformInfo info);
 class PageTransformerBuilder extends PageTransformer {
   final PageTransformerBuilderCallback builder;
 
-  PageTransformerBuilder({bool reverse: false, @required this.builder})
+  PageTransformerBuilder({bool reverse: false, required this.builder})
       : assert(builder != null),
         super(reverse: reverse);
 
@@ -101,11 +102,11 @@ class TransformerPageController extends PageController {
     bool keepPage = true,
     double viewportFraction = 1.0,
     this.loop: false,
-    this.itemCount,
+    required this.itemCount,
     this.reverse: false,
   }) : super(
             initialPage: TransformerPageController._getRealIndexFromRenderIndex(
-                initialPage ?? 0, loop, itemCount, reverse),
+                initialPage, loop, itemCount, reverse)!,
             keepPage: keepPage,
             viewportFraction: viewportFraction);
 
@@ -113,33 +114,33 @@ class TransformerPageController extends PageController {
     return _getRenderIndexFromRealIndex(index, loop, itemCount, reverse);
   }
 
-  int getRealItemCount() {
+  int? getRealItemCount() {
     if (itemCount == 0) return 0;
     return loop ? itemCount + kMaxValue : itemCount;
   }
 
   static _getRenderIndexFromRealIndex(
-      num index, bool loop, int itemCount, bool reverse) {
+      num? index, bool loop, int? itemCount, bool reverse) {
     if (itemCount == 0) return 0;
-    int renderIndex;
+    int? renderIndex;
     if (loop) {
-      renderIndex = index - kMiddleValue;
-      renderIndex = renderIndex % itemCount;
+      renderIndex = index! - kMiddleValue as int?;
+      renderIndex = renderIndex! % itemCount!;
       if (renderIndex < 0) {
         renderIndex += itemCount;
       }
     } else {
-      renderIndex = index;
+      renderIndex = index as int?;
     }
     if (reverse) {
-      renderIndex = itemCount - renderIndex - 1;
+      renderIndex = itemCount! - renderIndex! - 1;
     }
 
     return renderIndex;
   }
 
-  double get realPage {
-    double page;
+  double? get realPage {
+    double? page;
     if (position.maxScrollExtent == null || position.minScrollExtent == null) {
       page = 0.0;
     } else {
@@ -150,11 +151,11 @@ class TransformerPageController extends PageController {
   }
 
   static _getRenderPageFromRealPage(
-      double page, bool loop, int itemCount, bool reverse) {
-    double renderPage;
+      double? page, bool loop, int? itemCount, bool reverse) {
+    double? renderPage;
     if (loop) {
-      renderPage = page - kMiddleValue;
-      renderPage = renderPage % itemCount;
+      renderPage = page! - kMiddleValue;
+      renderPage = renderPage % itemCount!;
       if (renderPage < 0) {
         renderPage += itemCount;
       }
@@ -162,25 +163,25 @@ class TransformerPageController extends PageController {
       renderPage = page;
     }
     if (reverse) {
-      renderPage = itemCount - renderPage - 1;
+      renderPage = itemCount! - renderPage! - 1;
     }
 
     return renderPage;
   }
 
-  double get page {
+  double? get page {
     return loop
         ? _getRenderPageFromRealPage(realPage, loop, itemCount, reverse)
         : realPage;
   }
 
-  int getRealIndexFromRenderIndex(num index) {
+  int? getRealIndexFromRenderIndex(int index) {
     return _getRealIndexFromRenderIndex(index, loop, itemCount, reverse);
   }
 
-  static int _getRealIndexFromRenderIndex(
-      num index, bool loop, int itemCount, bool reverse) {
-    int result = reverse ? (itemCount - index - 1) : index;
+  static int? _getRealIndexFromRenderIndex(
+      int index, bool loop, int itemCount, bool reverse) {
+    int result = (reverse ? itemCount - index - 1 : index);
     if (loop) {
       result += kMiddleValue;
     }
@@ -192,7 +193,7 @@ class TransformerPageView extends StatefulWidget {
   /// Create a `transformed` widget base on the widget that has been passed to  the [PageTransformer.transform].
   /// See [TransformInfo]
   ///
-  final PageTransformer transformer;
+  final PageTransformer? transformer;
 
   /// Same as [PageView.scrollDirection]
   ///
@@ -200,7 +201,7 @@ class TransformerPageView extends StatefulWidget {
   final Axis scrollDirection;
 
   /// Same as [PageView.physics]
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// Set to false to disable page snapping, useful for custom scroll behavior.
   /// Same as [PageView.pageSnapping]
@@ -208,9 +209,9 @@ class TransformerPageView extends StatefulWidget {
 
   /// Called whenever the page in the center of the viewport changes.
   /// Same as [PageView.onPageChanged]
-  final ValueChanged<int> onPageChanged;
+  final ValueChanged<int>? onPageChanged;
 
-  final IndexedWidgetBuilder itemBuilder;
+  final IndexedWidgetBuilder? itemBuilder;
 
   // See [IndexController.mode],[IndexController.next],[IndexController.previous]
   final IndexController controller;
@@ -221,7 +222,7 @@ class TransformerPageView extends StatefulWidget {
   /// Animation curve
   final Curve curve;
 
-  final TransformerPageController pageController;
+  final TransformerPageController? pageController;
 
   /// Set true to open infinity loop mode.
   final bool loop;
@@ -248,9 +249,9 @@ class TransformerPageView extends StatefulWidget {
   /// [itemBuilder] will be called only with indices greater than or equal to
   /// zero and less than [itemCount].
   TransformerPageView({
-    Key key,
-    this.index,
-    Duration duration,
+    Key? key,
+    this.index = 0,
+    Duration? duration,
     this.curve: Curves.ease,
     this.viewportFraction: 1.0,
     this.loop: false,
@@ -258,34 +259,32 @@ class TransformerPageView extends StatefulWidget {
     this.physics,
     this.pageSnapping = true,
     this.onPageChanged,
-    this.controller,
+    required this.controller,
     this.transformer,
     this.itemBuilder,
     this.pageController,
-    @required this.itemCount,
-  })  : assert(itemCount != null),
-        assert(itemCount == 0 || itemBuilder != null || transformer != null),
+    required this.itemCount,
+  })   : assert(itemCount == 0 || itemBuilder != null || transformer != null),
         this.duration =
-            duration ?? new Duration(milliseconds: kDefaultTransactionDuration),
+            duration ?? Duration(milliseconds: kDefaultTransactionDuration),
         super(key: key);
 
   factory TransformerPageView.children(
-      {Key key,
-      int index,
-      Duration duration,
+      {Key? key,
+      required int index,
+      Duration? duration,
       Curve curve: Curves.ease,
       double viewportFraction: 1.0,
       bool loop: false,
       Axis scrollDirection = Axis.horizontal,
-      ScrollPhysics physics,
+      ScrollPhysics? physics,
       bool pageSnapping = true,
-      ValueChanged<int> onPageChanged,
-      IndexController controller,
-      PageTransformer transformer,
-      @required List<Widget> children,
-      TransformerPageController pageController}) {
-    assert(children != null);
-    return new TransformerPageView(
+      ValueChanged<int>? onPageChanged,
+      IndexController? controller,
+      PageTransformer? transformer,
+      required List<Widget> children,
+      TransformerPageController? pageController}) {
+    return TransformerPageView(
       itemCount: children.length,
       itemBuilder: (BuildContext context, int index) {
         return children[index];
@@ -301,18 +300,21 @@ class TransformerPageView extends StatefulWidget {
       scrollDirection: scrollDirection,
       physics: physics,
       onPageChanged: onPageChanged,
-      controller: controller,
+      controller: controller!,
     );
   }
 
   @override
   State<StatefulWidget> createState() {
-    return new _TransformerPageViewState();
+    return _TransformerPageViewState();
   }
 
-  static int getRealIndexFromRenderIndex(
-      {bool reverse, int index, int itemCount, bool loop}) {
-    int initPage = reverse ? (itemCount - index - 1) : index;
+  static int? getRealIndexFromRenderIndex(
+      {required bool reverse,
+      int index = 0,
+      int? itemCount,
+      required bool loop}) {
+    int? initPage = reverse ? (itemCount! - index - 1) : index;
     if (loop) {
       initPage += kMiddleValue;
     }
@@ -320,86 +322,88 @@ class TransformerPageView extends StatefulWidget {
   }
 
   static PageController createPageController(
-      {bool reverse,
-      int index,
-      int itemCount,
-      bool loop,
-      double viewportFraction}) {
-    return new PageController(
+      {required bool reverse,
+      required int index,
+      int? itemCount,
+      required bool loop,
+      required double viewportFraction}) {
+    return PageController(
         initialPage: getRealIndexFromRenderIndex(
-            reverse: reverse, index: index, itemCount: itemCount, loop: loop),
+            reverse: reverse, index: index, itemCount: itemCount, loop: loop)!,
         viewportFraction: viewportFraction);
   }
 }
 
 class _TransformerPageViewState extends State<TransformerPageView> {
-  Size _size;
-  int _activeIndex;
-  double _currentPixels;
+  Size? _size;
+  late int _activeIndex;
+  double? _currentPixels;
   bool _done = false;
 
   ///This value will not change until user end drag.
-  int _fromIndex;
+  late int _fromIndex;
 
-  PageTransformer _transformer;
+  PageTransformer? _transformer;
 
-  TransformerPageController _pageController;
+  TransformerPageController? _pageController;
 
   Widget _buildItemNormal(BuildContext context, int index) {
-    int renderIndex = _pageController.getRenderIndexFromRealIndex(index);
-    Widget child = widget.itemBuilder(context, renderIndex);
+    int renderIndex = _pageController!.getRenderIndexFromRealIndex(index);
+    Widget child = widget.itemBuilder!(context, renderIndex);
     return child;
   }
 
   Widget _buildItem(BuildContext context, int index) {
-    return new AnimatedBuilder(
-        animation: _pageController,
-        builder: (BuildContext c, Widget w) {
-          int renderIndex = _pageController.getRenderIndexFromRealIndex(index);
-          Widget child;
+    return AnimatedBuilder(
+        animation: _pageController!,
+        builder: (BuildContext c, Widget? w) {
+          int? renderIndex =
+              _pageController!.getRenderIndexFromRealIndex(index);
+          Widget? child;
           if (widget.itemBuilder != null) {
-            child = widget.itemBuilder(context, renderIndex);
+            child = widget.itemBuilder!(context, renderIndex);
           }
           if (child == null) {
-            child = new Container();
+            child = Container();
           }
           if (_size == null) {
-            return child ?? new Container();
+            return child;
           }
 
           double position;
 
-          double page = _pageController.realPage;
+          double? page = _pageController!.realPage;
 
-          if (_transformer.reverse) {
-            position = page - index;
+          if (_transformer!.reverse) {
+            position = page! - index;
           } else {
-            position = index - page;
+            position = index - page!;
           }
           position *= widget.viewportFraction;
 
-          TransformInfo info = new TransformInfo(
+          TransformInfo info = TransformInfo(
               index: renderIndex,
-              width: _size.width,
-              height: _size.height,
+              width: _size!.width,
+              height: _size!.height,
               position: position.clamp(-1.0, 1.0),
               activeIndex:
-                  _pageController.getRenderIndexFromRealIndex(_activeIndex),
+                  _pageController!.getRenderIndexFromRealIndex(_activeIndex),
               fromIndex: _fromIndex,
-              forward: _pageController.position.pixels - _currentPixels >= 0,
+              forward: _pageController!.position.pixels - _currentPixels! >= 0,
               done: _done,
               scrollDirection: widget.scrollDirection,
               viewportFraction: widget.viewportFraction);
-          return _transformer.transform(child, info);
+          return _transformer!.transform(child, info);
         });
   }
 
-  double _calcCurrentPixels() {
-    _currentPixels = _pageController.getRenderIndexFromRealIndex(_activeIndex) *
-        _pageController.position.viewportDimension *
-        widget.viewportFraction;
+  double? _calcCurrentPixels() {
+    _currentPixels =
+        _pageController!.getRenderIndexFromRealIndex(_activeIndex) *
+            _pageController!.position.viewportDimension *
+            widget.viewportFraction;
 
-    //  print("activeIndex:$_activeIndex , pix:$_currentPixels");
+    //  print('activeIndex:$_activeIndex , pix:$_currentPixels');
 
     return _currentPixels;
   }
@@ -408,20 +412,20 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   Widget build(BuildContext context) {
     IndexedWidgetBuilder builder =
         _transformer == null ? _buildItemNormal : _buildItem;
-    Widget child = new PageView.builder(
+    Widget child = PageView.builder(
       itemBuilder: builder,
-      itemCount: _pageController.getRealItemCount(),
+      itemCount: _pageController!.getRealItemCount(),
       onPageChanged: _onIndexChanged,
       controller: _pageController,
       scrollDirection: widget.scrollDirection,
       physics: widget.physics,
       pageSnapping: widget.pageSnapping,
-      reverse: _pageController.reverse,
+      reverse: _pageController!.reverse,
     );
     if (_transformer == null) {
       return child;
     }
-    return new NotificationListener(
+    return NotificationListener(
         onNotification: (ScrollNotification notification) {
           if (notification is ScrollStartNotification) {
             _calcCurrentPixels();
@@ -440,18 +444,16 @@ class _TransformerPageViewState extends State<TransformerPageView> {
 
   void _onIndexChanged(int index) {
     _activeIndex = index;
-    if (widget.onPageChanged != null) {
-      widget.onPageChanged(_pageController.getRenderIndexFromRealIndex(index));
-    }
+    widget.onPageChanged!(_pageController!.getRenderIndexFromRealIndex(index));
   }
 
   void _onGetSize(_) {
-    Size size;
+    Size? size;
     if (context == null) {
       onGetSize(size);
       return;
     }
-    RenderObject renderObject = context.findRenderObject();
+    RenderObject? renderObject = context.findRenderObject();
     if (renderObject != null) {
       Rect bounds = renderObject.paintBounds;
       if (bounds != null) {
@@ -462,7 +464,7 @@ class _TransformerPageViewState extends State<TransformerPageView> {
     onGetSize(size);
   }
 
-  void onGetSize(Size size) {
+  void onGetSize(Size? size) {
     if (mounted) {
       setState(() {
         _size = size;
@@ -476,20 +478,21 @@ class _TransformerPageViewState extends State<TransformerPageView> {
     //  int index = widget.index ?? 0;
     _pageController = widget.pageController;
     if (_pageController == null) {
-      _pageController = new TransformerPageController(
+      _pageController = TransformerPageController(
           initialPage: widget.index,
           itemCount: widget.itemCount,
           loop: widget.loop,
+          viewportFraction: widget.viewportFraction,
           reverse:
-              widget.transformer == null ? false : widget.transformer.reverse);
+              widget.transformer == null ? false : widget.transformer!.reverse);
     }
     // int initPage = _getRealIndexFromRenderIndex(index);
-    // _pageController = new PageController(initialPage: initPage,viewportFraction: widget.viewportFraction);
-    _fromIndex = _activeIndex = _pageController.initialPage;
+    // _pageController = PageController(initialPage: initPage,viewportFraction: widget.viewportFraction);
+    _fromIndex = _activeIndex = _pageController!.initialPage;
 
     _controller = getNotifier();
     if (_controller != null) {
-      _controller.addListener(onChangeNotifier);
+      _controller!.addListener(onChangeNotifier);
     }
     super.initState();
   }
@@ -497,41 +500,42 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   @override
   void didUpdateWidget(TransformerPageView oldWidget) {
     _transformer = widget.transformer;
-    int index = widget.index ?? 0;
+    int index = widget.index;
     bool created = false;
     if (_pageController != widget.pageController) {
       if (widget.pageController != null) {
         _pageController = widget.pageController;
       } else {
         created = true;
-        _pageController = new TransformerPageController(
+        _pageController = TransformerPageController(
+            viewportFraction: widget.viewportFraction,
             initialPage: widget.index,
             itemCount: widget.itemCount,
             loop: widget.loop,
             reverse: widget.transformer == null
                 ? false
-                : widget.transformer.reverse);
+                : widget.transformer!.reverse);
       }
     }
 
-    if (_pageController.getRenderIndexFromRealIndex(_activeIndex) != index) {
-      _fromIndex = _activeIndex = _pageController.initialPage;
+    if (_pageController!.getRenderIndexFromRealIndex(_activeIndex) != index) {
+      _fromIndex = _activeIndex = _pageController!.initialPage;
       if (!created) {
-        int initPage = _pageController.getRealIndexFromRenderIndex(index);
-        _pageController.animateToPage(initPage,
+        int initPage = _pageController!.getRealIndexFromRenderIndex(index)!;
+        _pageController!.animateToPage(initPage,
             duration: widget.duration, curve: widget.curve);
       }
     }
     if (_transformer != null)
-      WidgetsBinding.instance.addPostFrameCallback(_onGetSize);
+      WidgetsBinding.instance!.addPostFrameCallback(_onGetSize);
 
     if (_controller != getNotifier()) {
       if (_controller != null) {
-        _controller.removeListener(onChangeNotifier);
+        _controller!.removeListener(onChangeNotifier);
       }
       _controller = getNotifier();
       if (_controller != null) {
-        _controller.addListener(onChangeNotifier);
+        _controller!.addListener(onChangeNotifier);
       }
     }
     super.didUpdateWidget(oldWidget);
@@ -540,17 +544,17 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   @override
   void didChangeDependencies() {
     if (_transformer != null)
-      WidgetsBinding.instance.addPostFrameCallback(_onGetSize);
+      WidgetsBinding.instance!.addPostFrameCallback(_onGetSize);
     super.didChangeDependencies();
   }
 
-  ChangeNotifier getNotifier() {
+  ChangeNotifier? getNotifier() {
     return widget.controller;
   }
 
   int _calcNextIndex(bool next) {
     int currentIndex = _activeIndex;
-    if (_pageController.reverse) {
+    if (_pageController!.reverse) {
       if (next) {
         currentIndex--;
       } else {
@@ -564,11 +568,11 @@ class _TransformerPageViewState extends State<TransformerPageView> {
       }
     }
 
-    if (!_pageController.loop) {
-      if (currentIndex >= _pageController.itemCount) {
+    if (!_pageController!.loop) {
+      if (currentIndex >= _pageController!.itemCount) {
         currentIndex = 0;
       } else if (currentIndex < 0) {
-        currentIndex = _pageController.itemCount - 1;
+        currentIndex = _pageController!.itemCount - 1;
       }
     }
 
@@ -576,13 +580,13 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   }
 
   void onChangeNotifier() {
-    int event = widget.controller.event;
-    int index;
+    int? event = widget.controller.event;
+    int? index;
     switch (event) {
       case IndexController.MOVE:
         {
-          index = _pageController
-              .getRealIndexFromRenderIndex(widget.controller.index);
+          index = _pageController!
+              .getRealIndexFromRenderIndex(widget.controller.index!);
         }
         break;
       case IndexController.PREVIOUS:
@@ -595,23 +599,22 @@ class _TransformerPageViewState extends State<TransformerPageView> {
         //ignore this event
         return;
     }
-    if (widget.controller.animation) {
-      _pageController
-          .animateToPage(index,
-              duration: widget.duration, curve: widget.curve ?? Curves.ease)
+    if (widget.controller.animation!) {
+      _pageController!
+          .animateToPage(index!, duration: widget.duration, curve: widget.curve)
           .whenComplete(widget.controller.complete);
     } else {
-      _pageController.jumpToPage(index);
+      _pageController!.jumpToPage(index!);
       widget.controller.complete();
     }
   }
 
-  ChangeNotifier _controller;
+  ChangeNotifier? _controller;
 
   void dispose() {
     super.dispose();
     if (_controller != null) {
-      _controller.removeListener(onChangeNotifier);
+      _controller!.removeListener(onChangeNotifier);
     }
   }
 }
