@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_card_swipper/flutter_card_swiper.dart';
 
-void main() => runApp(new MyApp());
+void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
-      home: new InnerSwiper(),
+    return MaterialApp(
+      home: InnerSwiper(),
     );
   }
 }
@@ -15,7 +15,7 @@ class MyApp extends StatelessWidget {
 class InnerSwiper extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
-    return new _InnerSwiperState();
+    return _InnerSwiperState();
   }
 }
 
@@ -28,59 +28,60 @@ class _InnerSwiperState extends State<InnerSwiper> {
 
   @override
   void initState() {
-    controller = new SwiperController();
-    autoplayes = new List()
+    controller = SwiperController();
+    autoplayes = []
       ..length = 10
       ..fillRange(0, 10, false);
-    controllers = new List()
+    controllers = []
       ..length = 10
-      ..fillRange(0, 10, new SwiperController());
+      ..fillRange(0, 10, SwiperController());
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return new Scaffold(
-      body: new Swiper(
+    return Scaffold(
+      body: Swiper(
+        index: 0,
         loop: false,
         itemCount: 10,
         controller: controller,
-        pagination: new SwiperPagination(),
+        pagination: const SwiperPagination(),
         itemBuilder: (BuildContext context, int index) {
-          return new Column(
+          return Column(
             children: <Widget>[
-              new SizedBox(
-                child: new Swiper(
+              SizedBox(
+                child: Swiper(
                   controller: controllers[index],
-                  pagination: new SwiperPagination(),
+                  pagination: SwiperPagination(),
                   itemCount: 4,
                   itemBuilder: (BuildContext context, int index) {
-                    return new Container(
+                    return Container(
                       color: Colors.greenAccent,
-                      child: new Text("jkfjkldsfjd"),
+                      child: Text('jkfjkldsfjd'),
                     );
                   },
                   autoplay: autoplayes[index],
                 ),
                 height: 300.0,
               ),
-              new RaisedButton(
+              RaisedButton(
                 onPressed: () {
                   setState(() {
                     autoplayes[index] = true;
                   });
                 },
-                child: new Text("Start autoplay"),
+                child: const Text('Start autoplay'),
               ),
-              new RaisedButton(
+              RaisedButton(
                 onPressed: () {
                   setState(() {
                     autoplayes[index] = false;
                   });
                 },
-                child: new Text("End autoplay"),
+                child: Text('End autoplay'),
               ),
-              new Text("is autoplay: ${autoplayes[index]}")
+              Text('is autoplay: ${autoplayes[index]}')
             ],
           );
         },
