@@ -1,5 +1,5 @@
 const List<String> images = [
-  "images/bg0.jpeg",
-  "images/bg1.jpeg",
-  "images/bg2.jpeg",
+  'images/bg0.jpeg',
+  'images/bg1.jpeg',
+  'images/bg2.jpeg',
 ];
