@@ -10,14 +10,13 @@ class FormWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return new Padding(
-        padding: new EdgeInsets.all(5.0),
-        child: new Row(
+    return Padding(
+        padding: EdgeInsets.all(5.0),
+        child: Row(
           children: <Widget>[
-            new Text(label, style: new TextStyle(fontSize: 14.0)),
-            new Expanded(
-                child:
-                    new Align(alignment: Alignment.centerRight, child: child))
+            Text(label, style: const TextStyle(fontSize: 14.0)),
+            Expanded(
+                child: Align(alignment: Alignment.centerRight, child: child))
           ],
         ));
   }
@@ -57,9 +56,9 @@ class _FormSelectState extends State<FormSelect> {
     String placeholder = widget.placeholder;
     List<dynamic> values = widget.values;
 
-    return new Container(
-      child: new InkWell(
-        child: new Text(_selectedIndex < 0
+    return Container(
+      child: InkWell(
+        child: Text(_selectedIndex < 0
             ? placeholder
             : values[_selectedIndex].toString()),
         onTap: () {
@@ -67,25 +66,25 @@ class _FormSelectState extends State<FormSelect> {
           showBottomSheet(
               context: context,
               builder: (BuildContext context) {
-                return new SizedBox(
+                return SizedBox(
                   height: values.length * 30.0 + 200.0,
-                  child: new Column(
+                  child: Column(
                     mainAxisSize: MainAxisSize.min,
                     children: <Widget>[
-                      new SizedBox(
+                      SizedBox(
                         height: values.length * 30.0 + 70.0,
-                        child: new CupertinoPicker(
+                        child: CupertinoPicker(
                           itemExtent: 30.0,
                           children: values.map((dynamic value) {
-                            return new Text(value.toString());
+                            return Text(value.toString());
                           }).toList(),
                           onSelectedItemChanged: (int index) {
                             _selectedIndex = index;
                           },
                         ),
                       ),
-                      new Center(
-                        child: new RaisedButton(
+                      Center(
+                        child: RaisedButton(
                           onPressed: () {
                             if (_selectedIndex >= 0) {
                               widget
@@ -96,7 +95,7 @@ class _FormSelectState extends State<FormSelect> {
 
                             Navigator.of(context).pop();
                           },
-                          child: new Text("ok"),
+                          child: Text('ok'),
                         ),
                       )
                     ],
@@ -128,15 +127,15 @@ class NumberPad extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return new Row(
+    return Row(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
-        new IconButton(icon: new Icon(Icons.exposure_neg_1), onPressed: onSub),
-        new Text(
+        IconButton(icon: Icon(Icons.exposure_neg_1), onPressed: onSub),
+        Text(
           number is int ? number.toString() : number.toStringAsFixed(1),
-          style: new TextStyle(fontSize: 14.0),
+          style: TextStyle(fontSize: 14.0),
         ),
-        new IconButton(icon: new Icon(Icons.exposure_plus_1), onPressed: onAdd)
+        IconButton(icon: Icon(Icons.exposure_plus_1), onPressed: onAdd)
       ],
     );
   }
