@@ -6,19 +6,19 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   testWidgets('Control horizontal', (WidgetTester tester) async {
-    SwiperController controller = new SwiperController();
+    SwiperController controller = SwiperController();
 
-    SwiperPluginConfig config = new SwiperPluginConfig(
+    SwiperPluginConfig config = SwiperPluginConfig(
         activeIndex: 0,
         controller: controller,
         itemCount: 10,
         loop: true,
         scrollDirection: Axis.horizontal);
 
-    Key key = new UniqueKey();
-    await tester.pumpWidget(new MaterialApp(
-      home: new Scaffold(body: new Builder(builder: (BuildContext context) {
-        return new SwiperControl(key: key).build(context, config);
+    Key key = UniqueKey();
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(body: Builder(builder: (BuildContext context) {
+        return SwiperControl(key: key).build(context, config);
       })),
     ));
 
@@ -37,19 +37,19 @@ void main() {
   });
 
   testWidgets('Control vertical', (WidgetTester tester) async {
-    SwiperController controller = new SwiperController();
+    SwiperController controller = SwiperController();
 
-    SwiperPluginConfig config = new SwiperPluginConfig(
+    SwiperPluginConfig config = SwiperPluginConfig(
         activeIndex: 0,
         controller: controller,
         itemCount: 10,
         loop: true,
         scrollDirection: Axis.vertical);
 
-    Key key = new UniqueKey();
-    await tester.pumpWidget(new MaterialApp(
-      home: new Scaffold(body: new Builder(builder: (BuildContext context) {
-        return new SwiperControl(
+    Key key = UniqueKey();
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(body: Builder(builder: (BuildContext context) {
+        return SwiperControl(
                 key: key, color: Colors.white, disableColor: Colors.black87)
             .build(context, config);
       })),
